fix(backend): weight_kg was INT in DB/Go structs, truncating all sub-1kg demo readings to 0. Changed schema to DECIMAL(10,3) and Go structs/casts to float64 to preserve precision end-to-end

+++ pkg/handlers/telemetry_handler.go
@@ -24,7 +24,7 @@ func NewTelemetryHandler(db *sql.DB) *TelemetryHandler {
 // TelemetryDataPoint represents a single telemetry data point for charts
 type TelemetryDataPoint struct {
 	Timestamp      string   `json:"timestamp"`
-	WeightKg       int      `json:"weight_kg"`
+	WeightKg       float64  `json:"weight_kg"`
 	Latitude       float64  `json:"latitude"`
 	Longitude      float64  `json:"longitude"`
 	FuelLevel      *float64 `json:"fuel_level_liters,omitempty"`
